+++ src/parser/cpp.rs
@@ -103,7 +103,7 @@ struct CppContext<'a> {
 impl<'a> CppContext<'a> {
     fn new(
         source_code: &'a str,
-        lines: Vec<Vec<u8>>,
+        lines: Vec<&'a [u8]>,
         file_name: &'a str,
         tags: &'a mut Vec<tag::Tag>,
         tag_config: &'a TagKindConfig,