+++ src/parser/python.rs
@@ -61,7 +61,7 @@ struct PythonContext<'a> {
 impl<'a> PythonContext<'a> {
     fn new(
         source_code: &'a str,
-        lines: Vec<Vec<u8>>,
+        lines: Vec<&'a [u8]>,
         file_name: &'a str,
         tags: &'a mut Vec<tag::Tag>,
         tag_config: &'a TagKindConfig,