+++ src/parser/rust.rs
@@ -65,7 +65,7 @@ struct RustContext<'a> {
 impl<'a> RustContext<'a> {
     fn new(
         source_code: &'a str,
-        lines: Vec<Vec<u8>>,
+        lines: Vec<&'a [u8]>,
         file_name: &'a str,
         tags: &'a mut Vec<tag::Tag>,
         tag_config: &'a TagKindConfig,